// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const BookingStatus = {
  CONFIRMED: 'CONFIRMED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus]


export const UserRole = {
  USER: 'USER',
  ADMIN: 'ADMIN',
  STUDENT: 'STUDENT',
  TECHNICIAN: 'TECHNICIAN',
  TUTOR: 'TUTOR',
  MODERATOR: 'MODERATOR'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const AnalysisType = {
  ATS_SCAN: 'ATS_SCAN',
  JOB_MATCHER: 'JOB_MATCHER'
} as const

export type AnalysisType = (typeof AnalysisType)[keyof typeof AnalysisType]


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  BANNED: 'BANNED',
  DELETED: 'DELETED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  SUCCESS: 'SUCCESS',
  FAILED: 'FAILED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const PaymentMethod = {
  STRIPE: 'STRIPE',
  SSLCOMMERZ: 'SSLCOMMERZ',
  PAYPAL: 'PAYPAL'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const CreditType = {
  PURCHASE: 'PURCHASE',
  USAGE: 'USAGE',
  REFUND: 'REFUND',
  ADMIN_ADJUST: 'ADMIN_ADJUST'
} as const

export type CreditType = (typeof CreditType)[keyof typeof CreditType]


export const BlogStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  ARCHIVED: 'ARCHIVED'
} as const

export type BlogStatus = (typeof BlogStatus)[keyof typeof BlogStatus]


export const IssueType = {
  ISSUE: 'ISSUE',
  IMPROVEMENT: 'IMPROVEMENT',
  FEEDBACK: 'FEEDBACK'
} as const

export type IssueType = (typeof IssueType)[keyof typeof IssueType]


export const IssueStatus = {
  PENDING: 'PENDING',
  SUCCESS: 'SUCCESS'
} as const

export type IssueStatus = (typeof IssueStatus)[keyof typeof IssueStatus]